[ci] Fix Mock QPU tests and ensure CI runs them
