+++ python/tests/backends/test_IonQ.py
@@ -13,13 +13,15 @@ from typing import List
 from multiprocessing import Process
 from network_utils import check_server_connection
 try:
-    from utils.mock_qpu.ionq import startServer
-except:
+    from utils.mock_qpu import start_server, get_backend_port
+except Exception as e:
+    if os.environ.get("MOCK_NO_SKIP", "").lower() == "true":
+        raise e
     print("Mock qpu not available, skipping IonQ tests.")
     pytest.skip("Mock qpu not available.", allow_module_level=True)
 
-# Define the port for the mock server
-port = 62441
+TARGET_QPU = "ionq"
+port = get_backend_port(TARGET_QPU)
 
 
 def assert_close(got) -> bool:
@@ -31,10 +33,10 @@ def startUpMockServer():
     os.environ["IONQ_API_KEY"] = "00000000000000000000000000000000"
 
     # Set the targeted QPU
-    cudaq.set_target("ionq", url="http://localhost:{}".format(port))
+    cudaq.set_target(TARGET_QPU, url="http://localhost:{}".format(port))
 
     # Launch the Mock Server
-    p = Process(target=startServer, args=(port,))
+    p = Process(target=start_server, args=(TARGET_QPU,))
     p.start()
 
     if not check_server_connection(port):
