+++ python/tests/backends/test_OQC.py
@@ -17,13 +17,15 @@ from cudaq import spin
 import numpy as np
 
 try:
-    from utils.mock_qpu.oqc import startServer
-except:
+    from utils.mock_qpu import start_server, get_backend_port
+except Exception as e:
+    if os.environ.get("MOCK_NO_SKIP", "").lower() == "true":
+        raise e
     print("Mock qpu not available, skipping OQC tests.")
     pytest.skip("Mock qpu not available.", allow_module_level=True)
 
-# Define the port for the mock server
-port = 62442
+TARGET_QPU = "oqc"
+port = get_backend_port(TARGET_QPU)
 
 
 def assert_close(got) -> bool:
@@ -38,7 +40,7 @@ def startUpMockServer():
     os.environ["OQC_URL"] = f"http://localhost:{port}"
 
     # Launch the Mock Server
-    p = Process(target=startServer, args=(port,))
+    p = Process(target=start_server, args=(TARGET_QPU,))
     p.start()
 
     if not check_server_connection(port):
@@ -47,7 +49,7 @@ def startUpMockServer():
                     returncode=1)
 
     # Set the targeted QPU
-    cudaq.set_target('oqc',
+    cudaq.set_target(TARGET_QPU,
                      url=f'http://localhost:{port}',
                      auth_token="fake_auth_token")
     yield "Running the tests."
